requirements.txt
Django==3.2.4
graphene-django==3.0.0
